requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
cfgv==3.4.0
channels==4.2.0
channels_redis==4.2.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.3
django-auditlog==3.0.0
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-filter==24.3
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
filelock==3.16.1
h11==0.14.0
identify==2.6.3
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
msgpack==1.1.0
nodeenv==1.9.1
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.22.0
schema==0.7.7
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
uvicorn==0.32.1
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
websockets==14.1
